+++ test/test_hash_table.cpp
@@ -13,7 +13,7 @@ TEST(TestHashtable, TestInsert) {
   parlay::hashtable<parlay::hash_numeric<int>>
     table(400000, parlay::hash_numeric<int>{});;
   
-  parlay::parallel_for(1, 100000, [&](size_t i) {
+  parlay::parallel_for(1, 100000, [&](int i) {
     table.insert(i);
   });
 }
@@ -22,11 +22,11 @@ TEST(TestHashtable, TestFind) {
   parlay::hashtable<parlay::hash_numeric<int>>
     table(400000, parlay::hash_numeric<int>{});;
   
-  parlay::parallel_for(1, 100000, [&](size_t i) {
+  parlay::parallel_for(1, 100000, [&](int i) {
     table.insert(i);
   });
   
-  parlay::parallel_for(1, 100000, [&](size_t i) {
+  parlay::parallel_for(1, 100000, [&](int i) {
     auto val = table.find(i);
     ASSERT_EQ(val, i);
   });
@@ -36,22 +36,22 @@ TEST(TestHashtable, TestDelete) {
   parlay::hashtable<parlay::hash_numeric<int>>
     table(400000, parlay::hash_numeric<int>{});;
   
-  parlay::parallel_for(1, 100000, [&](size_t i) {
+  parlay::parallel_for(1, 100000, [&](int i) {
     table.insert(i);
   });
   
-  parlay::parallel_for(1, 100000, [&](size_t i) {
+  parlay::parallel_for(1, 100000, [&](int i) {
     auto val = table.find(i);
     ASSERT_EQ(val, i);
   });
   
-  parlay::parallel_for(1, 100000, [&](size_t i) {
+  parlay::parallel_for(1, 100000, [&](int i) {
     if (i % 2 == 0) {
       table.deleteVal(i);
     }
   });
   
-  parlay::parallel_for(1, 100000, [&](size_t i) {
+  parlay::parallel_for(1, 100000, [&](int i) {
     auto val = table.find(i);
     if (i % 2 == 1) {
       ASSERT_EQ(val, i);
